+++ src/global/storage/storage.service.ts
@@ -1,6 +1,7 @@
 import { GetObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { Inject, Injectable } from "@nestjs/common";
+import "multer";
 import { v4 as uuid } from "uuid";
 
 @Injectable()